pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gamepie"
version = "0.1.0"
description = "Library to create 2D games, powered by pygame :: Gamepie YouTube: 'https://www.youtube.com/@gamepieengine'"
readme = "README.md"
authors = [
  { name = "Qller Studio" }
]
license = { text = "MIT" }
requires-python = ">=3.7"
dependencies = [
  "numpy",
  "pygame",
  "pillow"
]

classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS :: MacOS X",
  "License :: OSI Approved :: MIT License"
]

[project.scripts]
gphelp = "gamepie.cmds.help:main"
gp = "gamepie.cmds.abt:main"
gpabout = "gamepie.cmds.abt:main"
gpplug = "gamepie.cmds.plug:main"
gpbuild = "gamepie.cmds.bld:main"
gpupdater = "gamepie.cmds.upt:main"
gpinit = "gamepie.cmds.init:main"
gpgit = "gamepie.cmds.webs:opengit"
gpyt = "gamepie.cmds.webs:openyt"

[project.optional-dependencies]
build = ["pyinstaller","toml"]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"gamepie" = [
  "assets/*",
  "assets/**/*",
  "gui/*",
  "gui/**/*",
  "input/*",
  "input/**/*",
  "load/*",
  "load/**/*",
  "draw/*",
  "draw/**/*",
  "plugins/*",
  "plugins/**/*"
]
